Reorganise projet structure
- Added a scripts/ folder
- Moved royalrecap.js, background.js and defaults.js to scripts/
- Added a path alias for the extension types
