+++ manifest.json
@@ -10,11 +10,11 @@
     "content_scripts": [
         {
             "matches": ["*://*.royalroad.com/*"],
-            "js": ["royalrecap.js"]
+            "js": ["scripts/royalrecap.js"]
         }
     ],
     "background": {
-        "scripts": ["background.js"],
+        "scripts": ["scripts/background.js"],
         "type": "module"
     },
     "options_ui": {
