+++ scripts/background.js
@@ -1,8 +1,6 @@
 import DEFAULTS from "./defaults.js"
 
-/**
- * Saves the default values to the browser storage after the extension has been installed
- */
+// Saves the default values to the browser storage after the extension has been installed
 browser.runtime.onInstalled.addListener((details) => {
     if (details.reason === "install") {
         browser.storage.sync.set(DEFAULTS)
